requirements.txt
# Core HTTP and SharePoint
requests>=2.31.0
Office365-REST-Python-Client>=2.5.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.0.0

# Document Processing
openpyxl>=3.1.0
python-docx>=1.0.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
python-dateutil>=2.8.0

# Testing (dev dependencies - optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
